+++ examples/Uncovered/InternalClasses.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Core;
+
+class CoreClass
+{
+    public function __construct()
+    {
+        new \Other\OtherClass();
+    }
+}
+
+namespace Other;
+
+class OtherClass {}