+++ src/RulesetEngine.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace SensioLabs\Deptrac;
 
+use JetBrains\PHPStormStub\PhpStormStubsMap;
+use SensioLabs\Deptrac\AstRunner\AstMap\ClassLikeName;
 use SensioLabs\Deptrac\Configuration\Configuration;
 use SensioLabs\Deptrac\Dependency\Result;
 use SensioLabs\Deptrac\RulesetEngine\Allowed;
@@ -33,7 +35,9 @@ class RulesetEngine
                 $layersNamesClassB = $classNameLayerResolver->getLayersByClassName($dependency->getClassLikeNameB());
 
                 if (0 === count($layersNamesClassB)) {
-                    $rules[] = new Uncovered($dependency, $layerName);
+                    if (!$this->ignoreUncoveredInternalClass($configuration, $dependency->getClassLikeNameB())) {
+                        $rules[] = new Uncovered($dependency, $layerName);
+                    }
                     continue;
                 }
 
@@ -58,5 +62,10 @@ class RulesetEngine
         }
 
         return new Context($rules);
+    }
+
+    private function ignoreUncoveredInternalClass(Configuration $configuration, ClassLikeName $classLikeName): bool
+    {
+        return $configuration->ignoreUncoveredInternalClasses() && isset(PhpStormStubsMap::CLASSES[$classLikeName->toString()]);
     }
 }