+++ tests/Collector/InheritsCollectorTest.php
@@ -6,23 +6,23 @@ namespace Tests\SensioLabs\Deptrac\Collector;
 
 use PHPUnit\Framework\TestCase;
 use SensioLabs\Deptrac\AstRunner\AstMap;
-use SensioLabs\Deptrac\Collector\ImplementsCollector;
+use SensioLabs\Deptrac\Collector\InheritsCollector;
 use SensioLabs\Deptrac\Collector\Registry;
 
-class ImplementsCollectorTest extends TestCase
+class InheritsCollectorTest extends TestCase
 {
     public function testGetType(): void
     {
-        static::assertEquals('implements', (new ImplementsCollector())->getType());
+        static::assertEquals('inherits', (new InheritsCollector())->getType());
     }
 
     public function dataProviderSatisfy(): iterable
     {
-        yield [['implements' => 'App\FizTrait'], false];
-        yield [['implements' => 'App\Bar'], true];
-        yield [['implements' => 'App\Baz'], true];
-        yield [['implements' => 'App\Foo'], false];
-        yield [['implements' => 'App\None'], false];
+        yield [['inherits' => 'App\FizTrait'], true];
+        yield [['inherits' => 'App\Bar'], true];
+        yield [['inherits' => 'App\Baz'], true];
+        yield [['inherits' => 'App\Foo'], true];
+        yield [['inherits' => 'App\None'], false];
     }
 
     /**
@@ -58,7 +58,7 @@ class ImplementsCollectorTest extends TestCase
             ->trait('App\FizTrait', 4);
         $fooBarFileReference = $fooBarFileReferenceBuilder->build();
 
-        $stat = (new ImplementsCollector())->satisfy(
+        $stat = (new InheritsCollector())->satisfy(
             $configuration,
             $fooBarFileReference->getAstClassReferences()[0],
             new AstMap([$fooFileReference, $barFileReference, $bazFileReference, $fooBarFileReference, $fizTraitFileReference]),